moved constants to color utils
