--- sites/theme-generator/src/lib/constants.ts
@@ -1,26 +1,0 @@
-export const colorNames = [
-  "primary",
-  "secondary",
-  "tertiary",
-  "success",
-  "warning",
-  "error",
-  "surface",
-] as const;
-
-export const colorShades = [
-  50, 100, 200, 300, 400, 500, 600, 700, 800, 900, 950,
-] as const;
-
-export const colorPairings = [
-  [50, 950],
-  [100, 900],
-  [200, 800],
-  [300, 700],
-  [400, 600],
-  [600, 400],
-  [700, 300],
-  [800, 200],
-  [900, 100],
-  [950, 50],
-] as const;
