+++ sites/theme-generator/src/lib/display/PreviewDisplay.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { colorNames, colorShades } from "$lib/constants";
+  import { colorNames, colorShades } from "$lib/utils/colors";
   import { bgColors, textColors } from "./colors";
   import { app } from "$lib/app.svelte";
   import PresetSection from "./PresetSection.svelte";
