+++ sites/theme-generator/src/lib/utils/colors.ts
@@ -1,5 +1,31 @@
 import chroma, { type Color } from "chroma-js";
-import { colorShades } from "$lib/constants";
+
+export const colorNames = [
+  "primary",
+  "secondary",
+  "tertiary",
+  "success",
+  "warning",
+  "error",
+  "surface",
+] as const;
+
+export const colorShades = [
+  50, 100, 200, 300, 400, 500, 600, 700, 800, 900, 950,
+] as const;
+
+export const colorPairings = [
+  [50, 950],
+  [100, 900],
+  [200, 800],
+  [300, 700],
+  [400, 600],
+  [600, 400],
+  [700, 300],
+  [800, 200],
+  [900, 100],
+  [950, 50],
+] as const;
 
 export function genScale(from: [string, string] | [string, string, string]) {
   const colors = chroma.scale(from).mode("oklch").colors(colorShades.length);
